added StatusCode attributes & exceptions handling

+++ OutOfSchool/OutOfSchool.WebApi/Controllers/ChildController.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using OutOfSchool.Services.Models;
-using OutOfSchool.WebApi.Extensions;
 using OutOfSchool.WebApi.Models;
 using OutOfSchool.WebApi.Services;
 
@@ -17,7 +19,7 @@ namespace OutOfSchool.WebApi.Controllers
     [Authorize(AuthenticationSchemes = "Bearer")]
     public class ChildController : ControllerBase
     {
-        private IChildService childService;
+        private readonly IChildService childService;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ChildController"/> class.
@@ -32,12 +34,11 @@ namespace OutOfSchool.WebApi.Controllers
         /// Get all children from the database.
         /// </summary>
         /// <returns>List of all children.</returns>
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Child>>> GetChildren()
         {
-            var children = await childService.GetAll().ConfigureAwait(false);
-            
-            return children.ToActionResult();
+            return Ok(await childService.GetAll().ConfigureAwait(false));
         }
 
         /// <summary>
@@ -45,12 +46,18 @@ namespace OutOfSchool.WebApi.Controllers
         /// </summary>
         /// <param name="id">The key in the database.</param>
         /// <returns>Child entity.</returns>
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpGet("{id}")]
         public async Task<ActionResult<ChildDTO>> GetChildById(long id)
         {
-            var child = await childService.GetById(id).ConfigureAwait(false);
+            if (id < 1 || childService.GetAll().Result.AsQueryable().Count() < id)
+            {
+                throw new ArgumentOutOfRangeException(id.ToString(), "The id is less than 1 or greater than number of table entities.");
+            }
 
-            return child.ToActionResult();
+            return Ok(await childService.GetById(id).ConfigureAwait(false));
         }
 
         /// <summary>
@@ -59,12 +66,20 @@ namespace OutOfSchool.WebApi.Controllers
         /// <param name="dto">Child entity to add.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
         [Authorize(Roles = "parent,admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPost]
         public async Task<ActionResult<Child>> CreateChild(ChildDTO dto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var child = await childService.Create(dto).ConfigureAwait(false);
 
-            return child.ToActionResult();
+            return Ok(child);
         }
 
         /// <summary>
@@ -73,12 +88,18 @@ namespace OutOfSchool.WebApi.Controllers
         /// <param name="dto">Child entity.</param>
         /// <returns>Child's key.</returns>
         [Authorize(Roles = "parent,admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpPut]
         public async Task<ActionResult> UpdateChild(ChildDTO dto)
         {
-            var child = await childService.Update(dto).ConfigureAwait(false);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
 
-            return child.ToActionResult();
+            return Ok(await childService.Update(dto).ConfigureAwait(false));
         }
 
         /// <summary>
@@ -87,12 +108,20 @@ namespace OutOfSchool.WebApi.Controllers
         /// <param name="id">Child's key.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
         [Authorize(Roles = "parent,admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteChild(long id)
         {
-            var childId = await childService.Delete(id).ConfigureAwait(false);
+            if (id < 1 || childService.GetAll().Result.AsQueryable().Count() < id)
+            {
+                throw new ArgumentOutOfRangeException(id.ToString(),
+                    "The id is less than 1 or greater than number of table entities.");
+            }
 
-            return childId.ToActionResult();
+            await childService.Delete(id).ConfigureAwait(false);
+
+            return Ok();
         }
     }
 }
